Updates to nested workflows on import

+++ web/util/workflow.js
@@ -8,7 +8,6 @@ const op = sequelize.Op;
 const stringSimilarity = require('string-similarity');
 const got = require('got');
 const fs = require('fs').promises;
-const { exit } = require('process');
 
 class Workflow {
 
@@ -80,7 +79,7 @@ class Workflow {
         try {
           let implementations = await models.implementation.findAll({where:{stepId:step.id}});
           for(let implementation of implementations) {
-            await fs.unlink("uploads/"+workflowId+"/"+implementation.language+"/"+implementation.fileName);
+            if(implementation.fileName.includes(".")) await fs.unlink("uploads/"+workflowId+"/"+implementation.language+"/"+implementation.fileName);
           } 
         } catch(exception) {
           console.error("Error deleting implementation:"+error);
@@ -178,7 +177,7 @@ class Workflow {
             if(!candidateChildStepInput) throw new Error(ERROR_PREFIX + "Error getting candidate child step input.");
             let candidateChildStepOutput=await models.output.findOne({where:{stepId:candidateChildStep.id}});
             if(!candidateChildStepOutput) throw new Error(ERROR_PREFIX + "Error getting candidate child step output.");
-            let workflowStepInput = await models.input.findAll({where:{stepId:workflowStep[0].id}})[0];
+            let workflowStepInput = await models.input.findOne({where:{stepId:workflowStep[0].id}});
             if(!workflowStepInput) throw new Error(ERROR_PREFIX + "Error getting workflow step input.");
             let workflowStepOutput = await models.output.findOne({where:{stepId:workflowStep[0].id}});
             if(!workflowStepOutput) throw new Error(ERROR_PREFIX + "Error getting workflow step output.");
